OnPing.DataServer

Description

OnPing Data Server

The OnPing Data Server stores homogeneously distributed values over time, using the binary store format.

Server run

runServer

Arguments

:: Int
Listening port

-> AwayServer
Away server to get old data from

-> IO ()

Run the OnPing Data Server on the given port.

Away servers

data AwayServer

A server to store unused values. It consists on a type of server, which specifies the way we communicate, its direction (IP or URL), and port.

Constructors

AwayServer AwayServerType String Int

data AwayServerType

Supported types of AwayServer.

Constructors

OnPingData
OnPing Data Server.

S3 String String
S3 Storage Server. It requires Key ID and Access Key.

DeadEnd
Dead End Server. It contains nothing. Use this when there is no server to connect.

deadEndServer :: AwayServer

A dead end away server with default configuration. It throws an error if you use it in the wrong way.