OnPing.DataServer.AwayServer

Description

Dynamic client interface. This is a low-level module. By all means, don't use it directly. Use OnPing.DataServer.Client instead.

Types

data AwayServerType

Supported types of AwayServer.

Constructors

OnPingData
OnPing Data Server.

S3 String String
S3 Storage Server. It requires Key ID and Access Key.

DeadEnd
Dead End Server. It contains nothing. Use this when there is no server to connect.

data AwayServer

A server to store unused values. It consists on a type of server, which specifies the way we communicate, its direction (IP or URL), and port.

Constructors

AwayServer AwayServerType String Int

deadEndServer :: AwayServer

A dead end away server with default configuration. It throws an error if you use it in the wrong way.

Client interface

clientTransfer

Arguments

:: AwayServer

-> Int
PID

-> Int
Initial time

-> IO TransferResponse

Ask the AwayServer to transfer a chunk of data.

data TransferResponse

Sum type of responses from a clientTransfer.

Constructors

TransferFound ByteString

TransferNotFound

TransferError String

clientDeleteChunk

Arguments

:: AwayServer

-> Int
PID

-> Int
Initial time

-> IO (Either String Bool)

Ask the AwayServer to delete a full chunk. Returns True if the chunk was successfully deleted.

clientDeleteChunkPlease

Arguments

:: AwayServer

-> Int
PID

-> Int
Initial time

-> Int
Maximum number of tries

-> IO (Either String Bool)

Ask the AwayServer to delete a full chunk repeatedly until the server reports that it was successfully removed, or the client gets tired of asking (maximum number of tries). Note that it won't try again if an error is received. Only if the request reached the server, and the server answered that the chunk could not be removed. As in clientDeleteChunk, it returns True when the chunk was successfully removed. If False is returned, it means the maximum number of tries was reached.

clientInsertChunk

Arguments

:: AwayServer

-> Int
PID

-> Int
Initial time

-> ByteString
Chunk to send

-> IO (Maybe String)

Send a chunk to AwayServer, and insert it at the given PID and initial time. If any error occurs in the process, a String is returned. Note that if the away server detects a valid header, it will accept the bytestring, even if it contains corrupted data. Make sure that the bytestring you are sending is an encoded binary store before calling this function.

When sending to DeadEnd, this function does nothing (and returns Nothing). Sending a chunk to oblivion is not supposed to be an error (or is it?).

Request primitives

generateFilePath

Arguments

:: Int
PID

-> Int
Initial time

-> FilePath

httpRequest

Arguments

:: String
Request URL or IP

-> Int
Port to make the request

-> String
Request route

-> RequestMethod
Request method

-> Maybe ByteString
Request content

-> IO (Either String (Response ByteString))
Response

Send an HTTP request, and return the response.

jsonRequest

Arguments

:: FromJSON a

=> String
Request URL or IP

-> Int
Port to make the request

-> String
Request route

-> RequestMethod
Request method

-> Maybe ByteString
Request content

-> IO (Either String a)
Parsed JSON response

Send an HTTP request and expect a JSON Value in return. If a non-JSON response is emitted, returns a parsing error message.