OnPing.DataServer.Client

Description

High-level client-side interface to the OnPing Data Server.

Away server type

data AwayServer

A server to store unused values. It consists on a type of server, which specifies the way we communicate, its direction (IP or URL), and port.

Constructors

AwayServer AwayServerType String Int

data AwayServerType

Supported types of AwayServer.

Constructors

OnPingData
OnPing Data Server.

S3 String String
S3 Storage Server. It requires Key ID and Access Key.

DeadEnd
Dead End Server. It contains nothing. Use this when there is no server to connect.

deadEndServer :: AwayServer

A dead end away server with default configuration. It throws an error if you use it in the wrong way.

Client monad

data Client m a

runClient :: Client m a -> AwayServer -> m a

lift :: MonadTrans t => forall m a. Monad m => m a -> t m a

Lift a computation from the argument monad to the constructed monad.

Client interface