OnPing.DataServer.Handler

Server global state

data ServerState

The server global state.

Constructors

ServerState

Fields

liveData :: IntMap (IntMap LiveValue)

IntMap of live data. The first key is the PID of an user. If a PID is not present, there is no live data of that user. You should look at the AwayServer. The second key (key of the inner map) refers to the initial time of the data in a binary store. Depending on the usage of the data, initial times will vary.

Live value maps should be manipulated strictly, while live values should be manipulated lazily.

awayServer :: AwayServer

Server to send the old data.

Types

type DecodedData = Decoded (Maybe Double)

Decoded data with holes.

type Counter = Int

The counter specifies the age of some live data. Live data is data kept in memory for quick access.

The counter has two limits:

First limit: After this limit is reached, live data is dropped from the global state, but the counter keeps growing. Therefore, the counter will stay in memory.

Second limit: After this limit, the file pointed by the Key is sent to S3. The counter is removed from memory as well.

data LiveValue

A live value in memory. It might or might not contain decoded data (lvData). If it does not contain the decoded data, it can be read from the local disk. The counter (lvCounter) tells you how much time this value has been on memory. See Counter.

Constructors

LiveValue

Fields

lvLength :: !Int

Length index (see lengthIndex) of decoded data at full resolution.

lvData :: Maybe DecodedData

Decoded data in memory.

lvCounter :: !Counter

Counter.

Yesod handler

The Handler monad. Server actions are performed here.

type Handler = HandlerT GlobalVar IO

data GlobalVar

Type of the global variable holding the server state.

createGlobalVar :: AwayServer -> IO GlobalVar

Manipulating the state

Primitive actions

get :: Handler ServerState

Read the current global state.

modify :: (ServerState -> ServerState) -> Handler ()

Modify the current state. The global state will be blocked from reading (see get) while this operation is performed.

Composed actions

addDecoded

Arguments

:: Int
PID

-> Int
Initial time

-> Int
Length index of decoded data at full resolution.

-> Maybe DecodedData
Decoded data.

-> Handler ()

Add DecodedData to memory (global state).

Request handlers

getTransferR

Arguments

:: Int
PID

-> Int
Initial time

-> Handler TypedContent

Send a chunk of data to the requester. It sends a 404 response if the chunk can't be found. If not found in the server, it will ask to the configured AwayServer. The 404 response contains an error description in its body. Otherwise (200 response), the body contains the chunk in binary format.

When the transference is accomplished, the response might be delayed by some seconds. This is because the server must make sure that - in case that the chunk was retrieved from the away server - the chunk no longer stays stored at the away server, thus avoiding conflicting copies to roll around. If, after a fixed number of tries, the file from the away server is not deleted, this server will stop trying and will report an error. However, even if an error is reported, the transferred data will still be included in the response body. A 200 response means that everything worked as expected. A 500 response means that the deletion on the away server failed. The details will be prompted in the server, and the HTTP response will contain the error in a X-Error header.

If a transference is done from an away server, but the data transferred was found corrupt, a 501 response will be sent, with the response body describing the error.

putDeleteChunkR

Arguments

:: Int
PID

-> Int
Initial time

-> Handler ()

Delete a chunk of data from the server, removing it from both memory and local disk. If it does not exist, this does nothing and reports success.

putInsertChunkR

Arguments

:: Int
PID

-> Int
Initial time

-> Handler ()

Insert a chunk received from a client request. It returns a 400 response if the header of the chunk is malformed. In that case, the body of the response will contain the error. Otherwise, a 200 response is sent.

Other Yesod Stuff

type Widget = WidgetT GlobalVar IO ()

resourcesGlobalVar :: [ResourceTree String]